+++ EssentialDeveloper/EssentialDeveloper/Feed API/FeedItemsMapper.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-internal final class FeedItemsMapper {
+public final class FeedItemsMapper {
     private struct Root: Decodable {
         private let items: [RemoteFeedItem]
         
@@ -19,7 +19,7 @@ internal final class FeedItemsMapper {
 
     private static var OK_200: Int { return 200}
     
-    internal static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
+    public static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
         guard response.statusCode == OK_200,
               let root = try? JSONDecoder().decode(Root.self, from: data)
         else {
