+++ EssentialDeveloper/EssentialDeveloperTests/Feed API/FeedItemsMapperTests.swift
@@ -1,0 +1,76 @@
+import XCTest
+import EssentialDeveloper
+
+final class FeedItemsMapperTests: XCTestCase {
+    func test_map_throwsErrorOnNon200HTTPResponse() throws {
+        let samples = [199, 201, 300, 400, 500]
+        let json = makeItemsJSON([])
+        
+        try samples.forEach { code in
+            XCTAssertThrowsError(
+                try FeedItemsMapper.map(json, from: HTTPURLResponse(statusCode: code))
+            )
+        }
+    }
+    
+    func test_map_throwsErrorOn200HTTPResponseWithInvalidJSON() throws {
+        let invalidJSON = Data("invalid json".utf8)
+        
+        XCTAssertThrowsError(
+            try FeedItemsMapper.map(invalidJSON, from: HTTPURLResponse(statusCode: 200))
+        )
+    }
+    
+    func test_map_deliverNoItemsOnHTTPResponseWithEmptyJSONList () throws {
+        let emptyListJSON = makeItemsJSON([])
+        
+        let result = try FeedItemsMapper.map(emptyListJSON, from: HTTPURLResponse(statusCode: 200))
+        XCTAssertEqual(result, [])
+    }
+    
+    // MARK: Happy Path
+    func test_map_deliversItemsOn200HTTPResponseWithJSONItems() throws {
+        let item1 = makeItem(
+            id: UUID(),
+            imageURL: URL(string:  "https://a-url.com")!
+        )
+        
+        let item2 = makeItem(
+            id: UUID(),
+            description: "A description",
+            location: "A location",
+            imageURL: URL(string:  "https://another-url.com")!
+        )
+        let json = makeItemsJSON([item1.json, item2.json])
+        
+        let result = try FeedItemsMapper.map(json, from: HTTPURLResponse(statusCode: 200))
+        XCTAssertEqual(result, [item1.model, item2.model])
+    }
+    
+    private func failure(_ error: RemoteFeedLoader.Error) -> RemoteFeedLoader.Result {
+        return .failure(error)
+    }
+    
+    private func makeItem(id: UUID, description: String? = nil, location: String? = nil, imageURL: URL) -> (model: FeedImage, json: [String: Any]) {
+        let item = FeedImage(id: id, description: description, location: location, url: imageURL)
+        let json = [
+            "id": id.uuidString,
+            "description": description,
+            "location": location,
+            "image": imageURL.absoluteString
+        ].compactMapValues { $0 }
+        
+        return (item, json)
+    }
+    
+    private func makeItemsJSON(_ items: [[String: Any]]) -> Data {
+        let itemsJSON = ["items": items]
+        return try! JSONSerialization.data(withJSONObject: itemsJSON)
+    }
+}
+
+private extension HTTPURLResponse {
+    convenience init(statusCode: Int) {
+        self.init(url: anyURL(), statusCode: statusCode, httpVersion: nil, headerFields: nil)!
+    }
+}
